refactor: extract tool execution routes and streaming helpers from http-server
Move createToolRouter (~400 lines), handleStreamingProxyCall,
handleStreamingToolCall, and sendSSEEvent into routes/tool-execution-routes.ts.
http-server.ts reduced from 1,338 to 789 lines (-41%).
